Replace old instances of platform with hook usage
